# docs/swagger.yaml
definitions:
...
  handlers.GetIssueHandlerData:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      issue:
        $ref: '#/definitions/models.Issue'
      project:
        $ref: '#/definitions/models.Project'
      tracker:
        $ref: '#/definitions/models.Tracker'
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  handlers.GetRolesHandlerData:
    properties:
      count:
        type: integer
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
...
  models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      issue_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Issue:
    properties:
      assigned_to_id:
        type: integer
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      project_id:
        type: integer
      status:
        type: string
      subject:
        type: string
      tracker_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Project:
    properties:
      created_on:
        type: string
      description:
        type: string
      id:
        type: integer
      identifier:
        type: string
      name:
        type: string
      parent_id:
        description: Usamos *int para manejar valores NULL
        type: integer
      updated_on:
        type: string
    type: object
  models.Role:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.Tracker:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password_hash:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: API para gestionar proyectos, usuarios, roles, categorías y problemas
    al estilo de Redmine.
  title: Go Redmine-ish API
  version: "1.0"
paths:
...
  /init:
    get:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
  /issue:
    post:
      parameters:
      - description: Issue
        in: body
        name: issue
        required: true
        schema:
          $ref: '#/definitions/models.Issue'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Issue'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
  /issue/{id}:
    delete:
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
    get:
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetIssueHandlerData'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
    put:
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: integer
      - description: Issue
        in: body
        name: issue
        required: true
        schema:
          $ref: '#/definitions/models.Issue'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Issue'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
...
  /role:
    post:
      parameters:
      - description: Role
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
  /role/{id}:
    delete:
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
    get:
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
  /roles:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetRolesHandlerData'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
...
securityDefinitions:
  BearerAuth:
    description: 'Ingresa el token en formato: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"